requirements.txt
# --- Core Deep Learning & Numerical Stack ---
# TensorFlow (Primary deep learning backend, includes Keras)
tensorflow==2.20.0

# NumPy (CRITICAL: Pinned to the version that avoided the 2.x conflict)
numpy==1.26.4 

# Scikit-learn (often used for data/metrics outside the model training loop)
scikit-learn==1.7.2

# SciPy (Used by NumPy/SciKit-learn/TensorFlow for scientific computing)
scipy==1.15.3

# --- Project-Specific Libraries ---
# Keras Utilities (for ImageDataGenerator, necessary due to older code structure)
tensorflow-addons==0.23.0 

# Image Processing (Used by your data generators and likely for pre/post-processing)
opencv-python==4.8.0.76

# Visualization (for Matplotlib, used in history plotting)
matplotlib==3.7.1

# Image Handling (Used by Matplotlib and data loaders)
Pillow==11.3.0

# --- Development & Jupyter Environment ---
# IPython Kernel (Makes the virtual environment selectable in Jupyter/VS Code)
ipykernel==6.30.1

Flask
gunicorn  # Recommended production web server

absl-py==2.3.1
altair==5.5.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.75.1
gunicorn==23.0.0
h5py==3.14.0
idna==3.11
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
narwhals==2.8.0
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.8.0.76
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.50.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-addons==0.23.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typeguard==2.13.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==1.17.3
absl-py==2.3.1
altair==5.5.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.75.1
gunicorn==23.0.0
h5py==3.14.0
idna==3.11
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
narwhals==2.8.0
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.8.0.76
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.50.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-addons==0.23.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typeguard==2.13.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==1.17.3
